test: restructure and add test to localtime
