+++ hermes/bin/src/runtime_extensions/hermes/localtime/host.rs
@@ -1,16 +1,11 @@
 //! Localtime host implementation for WASM runtime.
 
-use chrono::{Local, TimeZone};
-use chrono_tz::Tz;
-
+use super::localtime::{alt_localtime, get_localtime};
 use crate::{
     runtime_context::HermesRuntimeContext,
-    runtime_extensions::{
-        bindings::{
-            hermes::localtime::api::{Errno, Host, Localtime, Timezone},
-            wasi::clocks::wall_clock::Datetime,
-        },
-        hermes::localtime::get_tz,
+    runtime_extensions::bindings::{
+        hermes::localtime::api::{Errno, Host, Localtime, Timezone},
+        wasi::clocks::wall_clock::Datetime,
     },
 };
 
@@ -31,21 +26,7 @@ impl Host for HermesRuntimeContext {
     fn get_localtime(
         &mut self, when: Option<Datetime>, tz: Option<Timezone>,
     ) -> wasmtime::Result<Result<Localtime, Errno>> {
-        let timezone = get_tz(tz)?;
-        let local_naive = match when {
-            Some(Datetime {
-                seconds,
-                nanoseconds,
-            }) => {
-                let seconds = seconds.try_into().map_err(|_| Errno::InvalidLocaltime)?;
-                let utc_dt = chrono::DateTime::from_timestamp(seconds, nanoseconds)
-                    .ok_or(Errno::InvalidLocaltime)?;
-                utc_dt.naive_utc()
-            },
-            None => Local::now().naive_utc(),
-        };
-        let local_date_time = timezone.from_utc_datetime(&local_naive);
-        Ok(local_date_time.try_into())
+        Ok(get_localtime(when, tz))
     }
 
     /// Get a new localtime from a localtime, by recalculating time for a new timezone.
@@ -63,15 +44,7 @@ impl Host for HermesRuntimeContext {
     fn alt_localtime(
         &mut self, time: Localtime, tz: Option<Timezone>,
     ) -> wasmtime::Result<Result<Localtime, Errno>> {
-        let local_date_time: chrono::DateTime<Tz> = time.try_into()?;
-        let alt_local_date_time = match tz {
-            Some(alt_tz) => {
-                let tz = get_tz(Some(alt_tz))?;
-                tz.from_utc_datetime(&local_date_time.naive_utc())
-            },
-            None => local_date_time,
-        };
-        Ok(alt_local_date_time.try_into())
+        Ok(alt_localtime(time, tz))
     }
 
     /// Get a datetime from a localtime.
