+++ hermes/bin/src/runtime_extensions/hermes/localtime/mod.rs
@@ -9,6 +9,7 @@ use crate::runtime_extensions::bindings::{
 };
 
 mod host;
+mod localtime;
 
 /// Advise Runtime Extensions of a new context
 pub(crate) fn new_context(_ctx: &crate::runtime_context::HermesRuntimeContext) {}
